% Generated by roxygen2 (4.0.2): do not edit by hand
\name{binomialapproxLikelihood}
\alias{binomialapproxLikelihood}
\title{Calcuate the binomial approx likelihood from a set of observed, expected, and error values}
\usage{
binomialapproxLikelihood(observed, expected, errorValue, processError = 0,
  delta = 1e-11)
}
\arguments{
\item{observed}{The observed values for the likelihood}

\item{expected}{The expected values for the likelihood}

\item{errorValue}{The errorValues (variance estimates) for the likelihood}

\item{processError}{The additional process error for the likelihood}

\item{delta}{The robustifying constant to prevent divide by zero errors}
}
\description{
Calcuate the binomial approx likelihood from a set of observed, expected, and error values
}
\author{
Alistair Dunn
}

