% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirichletLikelihood.r
\name{dirichletLikelihood}
\alias{dirichletLikelihood}
\title{Calcuate the Dirichlet likelihood from a set of observed, expected, and error values}
\usage{
dirichletLikelihood(observed, expected, errorValue, processError = 0,
  delta = 1e-11)
}
\arguments{
\item{observed}{The observed proportions}

\item{expected}{The expected proportions}

\item{errorValue}{The error value (total sample size N) of the observed proportions}

\item{processError}{Additional process error for the likelihood (alpha0)}

\item{delta}{The robustifying constant to prevent divide by zero errors}
}
\description{
Calcuate the Dirichlet likelihood from a set of observed, expected, and error values
}
\author{
Alistair Dunn
}
