% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LognormalLikelihood.r
\name{lognormalLikelihood}
\alias{lognormalLikelihood}
\title{Calcuate the lognormal likelihood from a set of observed, expected, and error values}
\usage{
lognormalLikelihood(observed, expected, errorValue, processError = 0,
  delta = 1e-11)
}
\arguments{
\item{observed}{The observed values for the likelihood}

\item{expected}{The expected values for the likelihood}

\item{errorValue}{The errorValues (variance estimates) for the likelihood}

\item{processError}{The additional process error for the likelihood}

\item{delta}{The robustifying constant to prevent divide by zero errors}
}
\description{
Calcuate the lognormal likelihood from a set of observed, expected, and error values
}
\author{
Alistair Dunn
}

